requirements.txt
asgiref==3.11.0
Django==5.0.1
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
et-xmlfile==2.0.0
numpy==2.2.1
openpyxl==3.1.2
pandas==2.1.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
Pillow==10.2.0
psycopg2-binary==2.9.9
